Fix sendKey broadcast and examples
