+++ README.md
@@ -8,6 +8,8 @@
 Requires CEC capable device (e.g. Raspberry Pi or USB-CEC adapter).<br>
 Additionally `cec-client` must be installed. On Raspbian it is included in cec-utils package.
 
+Controller scans devices on startup. It takes a while (scan is done async and result is returned in "ready" event).
+
 ### Usage Examples
 ```javascript
 var CecController = require('cec-controller');
@@ -66,15 +68,13 @@ cecCtl.on('error', console.error);
 
 function readyHandler(controller)
 {
-	controller.setActive().then(() =>
+	/* In this example dev1 is a satellite decoder */
+	controller.dev1.sendKey('up').then((success) =>
 	{
-		controller.dev0.sendKey('up').then((success) =>
-		{
-			if(success)
-				console.log('Successfully send "up" key to TV');
-			else
-				console.error('Could not send input key!');
-		});
+		if(success)
+			console.log('Successfully send "up" key to decoder');
+		else
+			console.error('Could not send input key!');
 	});
 }
 ```
